backend/requirements.txt
fastapi
uvicorn
boto3
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv